Corrected the path size calculation
