+++ lemin_files/ft_ants_to_paths.c
@@ -6,7 +6,7 @@
 /*   By: dsaripap <dsaripap@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/06/19 10:11:44 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/06/19 10:59:30 by dsaripap      ########   odam.nl         */
+/*   Updated: 2020/06/19 14:28:31 by dsaripap      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,12 +63,13 @@ int					ft_all_ants_in_same_path(t_ant_farm *ant_farm)
 
 	i = 1;
 	tmp = ant_farm->paths;
+	ft_printf("All ants in same path\n");
 	while (i <= ant_farm->ants)
 	{
 		ant = (t_ants *)ft_memalloc(sizeof(t_ants));
 		ant->ant_id = i;
 		ant->path = tmp;
-		ft_printf(ANSI_COLOR_BLUE"ant_id: %d in path_id: %d\n"ANSI_COLOR_RESET, ant->ant_id, ant->path->path_id);
+		// ft_printf(ANSI_COLOR_BLUE"ant_id: %d in path_id: %d\n"ANSI_COLOR_RESET, ant->ant_id, ant->path->path_id);
 		i += 1;
 	}
 	return (0);
@@ -117,21 +118,21 @@ int					ft_ants_to_paths(t_ant_farm *ant_farm)
 
 	ft_printf("\nAssigning Ants to Paths\n");
 	ft_printf("========================\n");
-	// ft_printf("Number of ants : %d \n", ant_farm->ants);
+	ft_printf("Number of ants : %d \n", ant_farm->ants);
 	// ft_printf("Max paths : %d \n", ant_farm->max_paths);
-	// ft_printf("Discovered paths : %d \n", ant_farm->discovered_paths);
+	ft_printf("Discovered paths : %d \n", ant_farm->discovered_paths);
 	// ft_print_paths_list(ant_farm);
 	if (ant_farm->ants == 0)
-		return (0);
+		return (SUCCESS);
 	if (ant_farm->discovered_paths == 0)
-		return (0);
+		return (SUCCESS);
 	if (ant_farm->discovered_paths == 1)
 	{
 		ft_all_ants_in_same_path(ant_farm);
-		return (0);
+		return (SUCCESS);
 	}
 	i = 1;
-	while ((i <= ant_farm->ants) && (ant_farm->discovered_paths > 1))
+	while (i <= ant_farm->ants) //&& (ant_farm->discovered_paths > 1))
 	{
 		tmp = ant_farm->paths;
 		tmp_lst = tmp->path_lst->next;
@@ -144,13 +145,13 @@ int					ft_ants_to_paths(t_ant_farm *ant_farm)
 				next = cur->next;
 			else
 				next = ant_farm->paths;
-			// ft_printf(ANSI_COLOR_BLUE"cur_pathid: %d, size = %d\n"ANSI_COLOR_RESET, cur->path_id, cur->path_size + cur->ants_amount);
+			ft_printf(ANSI_COLOR_BLUE"cur_pathid: %d, size = %d\n"ANSI_COLOR_RESET, cur->path_id, cur->path_size + cur->ants_amount);
 			if (tmp->next != NULL)
-				;
-				// ft_printf(ANSI_COLOR_YELLOW_PAST"next_pathid: %d, size = %d\n"ANSI_COLOR_RESET, next->path_id, next->path_size + next->ants_amount);
+				// ;
+				ft_printf(ANSI_COLOR_YELLOW_PAST"next_pathid: %d, size = %d\n"ANSI_COLOR_RESET, next->path_id, next->path_size + next->ants_amount);
 			else
-				;
-				// ft_printf(ANSI_COLOR_YELLOW_PAST"next_pathid: %d, size = %d\n"ANSI_COLOR_RESET, ant_farm->paths->path_id, ant_farm->paths->path_size + ant_farm->paths->ants_amount);
+				// ;
+				ft_printf(ANSI_COLOR_YELLOW_PAST"next_pathid: %d, size = %d\n"ANSI_COLOR_RESET, ant_farm->paths->path_id, ant_farm->paths->path_size + ant_farm->paths->ants_amount);
 			if (cur->path_size + cur->ants_amount <= next->path_size + next->ants_amount)
 			{
 				// assign to current path
@@ -166,12 +167,11 @@ int					ft_ants_to_paths(t_ant_farm *ant_farm)
 			}
 			// ft_ants_addend(&(tmp_lst->room->ants), ant);
 			ft_ants_to_pathlst_addend(&(ant->path->ants_lst), ant);
-			// ft_printf(ANSI_COLOR_GREEN_EMER"ant_id: %d, path_id = %d\n\n"ANSI_COLOR_RESET, ant->ant_id, ant->path->path_id);
+			ft_printf(ANSI_COLOR_GREEN_EMER"ant_id: %d, path_id = %d\n\n"ANSI_COLOR_RESET, ant->ant_id, ant->path->path_id);
 			i += 1;
 		}
 	}
-	ft_print_ants(ant_farm);
-	ft_move_ants(ant_farm);
+	ft_print_ants_in_paths(ant_farm);
 	// ft_keep_ants_moving(ant_farm);
 	return (0);
 }
